+++ src/test/java/es/upm/grise/profundizacion/HandleDocuments/SmokeTest.java
@@ -13,7 +13,7 @@ public class SmokeTest {
 	@Test
 	public void formatTemplateCorrectly() throws NonRecoverableError, RecoverableError {
 		
-		Document d = new Document();
+		Document d = new Document(1623);
 		d.setTemplate("DECLARATION");
 		d.setTitle("A");
 		d.setAuthor("B");