+++ changes.md
@@ -1,1 +1,7 @@
 # Changes made to the code
+
+- DocumentID --> protected
+- Constructor por parámetros en Document 
+- DocumentIDProvider --> public
+- getDocumentId --> añadir parámetro instanceDocumentId
+- SmokeTest --> id por parámetros